Kick and Samplers outputs mixed; added decay control to kick

+++ Punck/src/samples.h
@@ -46,7 +46,7 @@ public:
 //	std::array<Bank, 10> bankA;				// Store pointer to Bank A/B samples sorted by index
 //	std::array<Bank, 10> bankB;
 
-	int32_t mixedSamples[2] = {};			// Left/right samples mixed and ready to output to DAC
+	float mixedSamples[2] = {};				// Left/right samples mixed and ready to output to DAC
 
 	Samples();
 	void Play(SamplePlayer s, uint32_t noteOffset, uint32_t noteRange);
